refactor(ruby): extract InstrumentationLogFilter; capture install failures correctly
- Move the log-suppression filter into its own file
  (instrumentation_log_filter.rb) with class-level capture_failures/
  failure_warning helpers (keeps OpenTelemetryConfig under the class-length
  limit and is unit-testable).
- Wrap the whole OpenTelemetry::SDK.configure call rather than just use_all:
  the SDK installs instrumentations AFTER the configure block returns, so the
  earlier placement around use_all never saw the install logging. Now the
  per-instrumentation 'failed to install' / 'successfully installed' chatter is
  actually suppressed and failures are collected for the single summary.
- rails.rb: Railtie#otel_logger_provider_available? now returns an explicit
  boolean instead of nil when the logs constant is absent. Fixes a pre-existing
  order-dependent failure in rails_railtie_test (verified failing identically on
  the unmodified gem at the same seed).

Updated e2e Gemfile.locks for the new individual-gem resolution. demo (7.2) and
api-only (7.2) suites pass; demo-rails70 (7.0) passes with 0 'failed to install'.
